# 方案/契约/openapi_2026-07-31.yaml
openapi: "3.0.3"
info:
  title: 南塘云村 API 契约
  description: |
    UI-FIX-P2-BE + BE补 完整契约——一营 FE 按此接入。
    含 B2-B4 储物管理 + B5 档案室 + B6 田间管理 + B7 用户设置 + B8 校核。
  version: "2026-07-31"
  contact:
    name: 南塘云村施工二营

servers:
  - url: /api
    description: 当前部署

tags:
  - name: storage
    description: 储物管理（B2-B4，本次新建）
  - name: archive
    description: 档案室（B5，已建）
  - name: fields
    description: 田间管理（B6，已建）
  - name: settings
    description: 用户设置（B7，已建）
  - name: verifications
    description: 校核列表（B8，已有）

paths:
  # ════════════════════════════════════════
  # B2: POST /storage/items
  # ════════════════════════════════════════
  /storage/items:
    post:
      tags: [storage]
      summary: 新增储物记录（B2）
      description: 村民新增一条冰箱/储物间/共享储物
      operationId: createStorageItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemReq"
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateItemRes"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

    # ════════════════════════════════════════
    # B3: GET /storage/items
    # ════════════════════════════════════════
    get:
      tags: [storage]
      summary: 查本人储物（B3）
      description: 返回当前用户所有未过期储物，按 storage_location 分组
      operationId: listStorageItems
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 按 location 分组的储物列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListItemsRes"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ════════════════════════════════════════
  # B4: DELETE /storage/items/{id}
  # ════════════════════════════════════════
  /storage/items/{id}:
    delete:
      tags: [storage]
      summary: 删储物（B4）
      description: 本人或 admin 可删。返回 204 No Content
      operationId: deleteStorageItem
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 删除成功
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ════════════════════════════════════════
  # B5: GET /archive/items
  # ════════════════════════════════════════
  /archive/items:
    get:
      tags: [archive]
      summary: 档案条目列表（B5）
      description: |
        聚合 Journal + ActivityLog + Announcement + CovenantSignature。
        复合 ID 格式：{source}_{db_id}（如 journal_123）。
        category 可选：log / event / contract。
      operationId: listArchiveItems
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [log, event, contract, ""]
      responses:
        "200":
          description: 归档条目列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveListRes"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /archive/items/{item_id}:
    get:
      tags: [archive]
      summary: 档案条目详情（B5）
      description: 按复合 ID（{source}_{db_id}）查单条归档详情
      operationId: getArchiveItem
      security:
        - bearerAuth: []
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 归档详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveDetailRes"
        "404":
          $ref: "#/components/responses/NotFound"

  # ════════════════════════════════════════
  # B6: GET /fields + 动作端点
  # ════════════════════════════════════════
  /fields:
    get:
      tags: [fields]
      summary: 地块列表（B6）
      operationId: listFieldPlots
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 地块列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldListRes"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /fields/{plot_id}:
    get:
      tags: [fields]
      summary: 地块详情（B6）
      operationId: getFieldPlot
      security:
        - bearerAuth: []
      parameters:
        - name: plot_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 地块详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldDetailRes"
        "404":
          $ref: "#/components/responses/NotFound"

  /fields/{plot_id}/harvest:
    post:
      tags: [fields]
      summary: 收割地块（B6）
      description: 仅成熟地块可收割 → 回到休耕
      operationId: harvestFieldPlot
      security:
        - bearerAuth: []
      parameters:
        - name: plot_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 收割成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldDetailRes"
        "400":
          $ref: "#/components/responses/BadRequest"

  /fields/{plot_id}/water:
    post:
      tags: [fields]
      summary: 浇水（B6）
      description: 更新 watered_at，缺水→健康；休耕→400
      operationId: waterFieldPlot
      security:
        - bearerAuth: []
      parameters:
        - name: plot_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 浇水成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldDetailRes"
        "400":
          $ref: "#/components/responses/BadRequest"

  /fields/{plot_id}/fertilize:
    post:
      tags: [fields]
      summary: 施肥（B6）
      description: 更新 fertilized_at，缺肥→健康；休耕→400
      operationId: fertilizeFieldPlot
      security:
        - bearerAuth: []
      parameters:
        - name: plot_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 施肥成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldDetailRes"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ════════════════════════════════════════
  # B7: GET/PATCH /users/me/settings
  # ════════════════════════════════════════
  /users/me/settings:
    get:
      tags: [settings]
      summary: 读用户设置（B7）
      description: 无记录则返回默认值 {notification:true, theme:light, language:zh-CN}
      operationId: getUserSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 用户设置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsRes"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      tags: [settings]
      summary: 局部更新用户设置（B7）
      description: 只改传入的字段（null=不改）。theme 枚举 light/dark，language 枚举 zh-CN/en
      operationId: patchUserSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchSettingsReq"
      responses:
        "200":
          description: 更新后的设置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsRes"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ════════════════════════════════════════
  # B8: GET /data/verifications（已有）
  # ════════════════════════════════════════
  /data/verifications:
    get:
      tags: [verifications]
      summary: 校核列表（B8 · 已有）
      description: |
        已存在于 data.py L188。admin 看全部，非 admin 只看自己。
        注意：verifier_id 对应返回体中的 `verifier` 字段；
        task_id 不在顶层，需从 detail JSON 提取。
      operationId: listVerifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 校核记录列表（最多 50 条）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Verification"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT token（POST /api/auth/login 获取）

  schemas:
    # ── Storage ──
    CreateItemReq:
      type: object
      required: [item_name, category, quantity, storage_location]
      properties:
        item_name:
          type: string
          minLength: 1
          maxLength: 100
          example: "白菜"
        category:
          type: string
          enum: ["食物", "工具", "杂物"]
        quantity:
          type: integer
          minimum: 1
          default: 1
        storage_location:
          type: string
          enum: ["冰箱", "储物间", "共享"]
        expires_at:
          type: string
          nullable: true
          format: date-time
          description: ISO 8601 过期时间（null=不过期）

    StorageItem:
      type: object
      properties:
        id:
          type: string
          example: "si_user1_1722300000_1234"
        user_id:
          type: string
        item_name:
          type: string
        category:
          type: string
          enum: ["食物", "工具", "杂物"]
        quantity:
          type: integer
          minimum: 1
        storage_location:
          type: string
          enum: ["冰箱", "储物间", "共享"]
        added_at:
          type: string
          format: date-time
        expires_at:
          type: string
          nullable: true
          format: date-time

    CreateItemRes:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        item:
          $ref: "#/components/schemas/StorageItem"

    ListItemsRes:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        items:
          type: object
          properties:
            冰箱:
              type: array
              items:
                $ref: "#/components/schemas/StorageItem"
            储物间:
              type: array
              items:
                $ref: "#/components/schemas/StorageItem"
            共享:
              type: array
              items:
                $ref: "#/components/schemas/StorageItem"

    # ── Archive (B5) ──
    ArchiveItem:
      type: object
      properties:
        id:
          type: string
          example: "journal_123"
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [log, event, contract]
        created_at:
          type: string
        author:
          type: string
        source:
          type: string
          enum: [journal, activity_log, announcement, covenant_signature]

    ArchiveListRes:
      type: object
      properties:
        ok:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/ArchiveItem"
        total:
          type: integer

    ArchiveDetailRes:
      type: object
      properties:
        ok:
          type: boolean
        item:
          $ref: "#/components/schemas/ArchiveItem"

    # ── Fields (B6) ──
    FieldPlot:
      type: object
      properties:
        id:
          type: string
        plot_name:
          type: string
        crop_name:
          type: string
          nullable: true
        planted_at:
          type: string
          nullable: true
        harvest_at:
          type: string
          nullable: true
        stage:
          type: string
          enum: [休耕, 播种, 生长, 成熟, 收割]
        health:
          type: string
          enum: [健康, 缺水, 缺肥, 病虫害]
        watered_at:
          type: string
          nullable: true
        fertilized_at:
          type: string
          nullable: true
        harvested_by:
          type: string
          nullable: true
        planted_by:
          type: string
          nullable: true
        created_at:
          type: string

    FieldListRes:
      type: object
      properties:
        ok:
          type: boolean
        plots:
          type: array
          items:
            $ref: "#/components/schemas/FieldPlot"
        total:
          type: integer

    FieldDetailRes:
      type: object
      properties:
        ok:
          type: boolean
        plot:
          $ref: "#/components/schemas/FieldPlot"

    # ── Settings (B7) ──
    SettingsRes:
      type: object
      properties:
        ok:
          type: boolean
        settings:
          $ref: "#/components/schemas/UserSettings"

    UserSettings:
      type: object
      properties:
        notification:
          type: boolean
          default: true
        theme:
          type: string
          enum: [light, dark]
          default: light
        language:
          type: string
          enum: [zh-CN, en]
          default: zh-CN

    PatchSettingsReq:
      type: object
      properties:
        notification:
          type: boolean
          nullable: true
          description: null=不改
        theme:
          type: string
          nullable: true
          description: light/dark，null=不改
        language:
          type: string
          nullable: true
          description: zh-CN/en，null=不改

    # ── Verification (B8) ──
    Verification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          example: "task"
        doer:
          type: string
        action:
          type: string
        detail:
          type: object
          description: JSON 对象（含 task_id 等关联信息）
        nt_amount:
          type: integer
        verifier_reward:
          type: integer
        status:
          type: string
          enum: ["pending", "approved", "rejected"]
        verifier:
          type: string
          nullable: true
          description: 校核人 ID（= 卡面 verifier_id）
        verified_at:
          type: string
          nullable: true
          format: date-time
        reject_reason:
          type: string
          nullable: true
        retry_count:
          type: integer
        created_at:
          type: string
          format: date-time

  responses:
    BadRequest:
      description: 请求参数错误（400）
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    Unauthorized:
      description: 未认证（401）
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    Forbidden:
      description: 无权限（403）
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    NotFound:
      description: 资源不存在（404）
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    ValidationError:
      description: 参数校验失败（422）
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
              error:
                type: string
              fields:
                type: array
                items:
                  type: object